Moved the catalog handler to parent package
