+++ api/handler/response.go
@@ -1,4 +1,4 @@
-package catalog
+package handler
 
 type catalogIdResponse struct {
 	Id string `json:"id"`
